requirements.txt
Django>=3.2.0,<4.1.0
djangorestframework>=3.12.0,<3.14.0

flake8>=4.0.0,<4.1.0
django-cors-headers==3.11.0
PyYAML==6.0
uritemplate==4.1.1

